+++ src/store/selfTest.ts
